Refactor(frontend): reorganise settings feature utilities
