+++ frontend/src/functionality/Settings/utils/Api/createCalendarSubscription.js
@@ -1,4 +1,4 @@
-import { api } from '../../api.js';
+import { api } from '../../../../api.js';
 
 const CALENDAR_SUBSCRIPTIONS_ENDPOINT = '/api/calendar-subscriptions/';
