+++ frontend/src/functionality/Settings/utils/Helpers/downloadFile.js
