+++ frontend/src/functionality/Settings/utils/Helpers/formatSubscriptionLastSyncedAt.js
@@ -1,0 +1,9 @@
+function formatSubscriptionLastSyncedAt(lastSyncedAt) {
+    if (!lastSyncedAt) {
+        return 'Never';
+    }
+
+    return new Date(lastSyncedAt).toLocaleString();
+}
+
+export default formatSubscriptionLastSyncedAt;
