+++ frontend/src/functionality/Settings/utils/Helpers/handleDeleteSubscription.js
