+++ frontend/src/functionality/Settings/utils/Helpers/handleRefreshSubscription.js
